% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample.R
\name{drsvp}
\alias{drsvp}
\title{Daily realized semivariance of a Portfolio Function}
\usage{
drsvp(data, weights = NULL)
}
\arguments{
\item{data}{A xts object containing the intraday prices
of the portfolio's assets.}

\item{weights}{A numeric vector containing the weights of each
asset in the portfolio. Default = NULL. If NULL, an equally weighted
portfolio is assumed.}
}
\value{
An xts object. The output has the following columns:
\itemize{
\item PSV : The daily realized semicovariance of the portfolio
\item NSV : The daily upside realized semicovariance of the portfolio
}
}
\description{
Return the Daily realized semivariance of a Portfolio
}
\examples{
# Generate a xts with the realized measures
realized <- drsvp(data = xts_data)
}
