% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample.R
\name{drvp}
\alias{drvp}
\title{Daily realized variance of a Portfolio Function}
\usage{
drvp(data, weights = NULL)
}
\arguments{
\item{data}{A xts object containing the intraday prices
of the portfolio's assets.}

\item{weights}{A numeric vector containing the weights of each
asset in the portfolio. Default = NULL. If NULL, an equally weighted
portfolio is assumed.}
}
\value{
An xts object. The output has the following columns:
\itemize{
\item RC : The daily realized semicovariance of the portfolio
\item P : The daily upside realized semicovariance of the portfolio
\item N : The daily downside realized semicovariance of the portfolio
\item M : The daily mixed realized semicovariance of the portfolio
}
}
\description{
Return the Daily realized variance of a Portfolio
}
\examples{
# Generate a xts with the realized measures
realized <- drvp(data = xts_data)
}
