% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample.R
\name{HAR_frame}
\alias{HAR_frame}
\title{HAR_frame Function}
\usage{
HAR_frame(ts = NULL, periods = c(1, 5, 22))
}
\arguments{
\item{ts}{A xts object containing the realized measures}

\item{periods}{A numeric vector containing the periods. Default 'c(1,5,22)'}
}
\value{
An xts object containing the response variable e covariates of a HAR.
Each covariate X_i has three columns:
\itemize{
\item X_i_d: The lagged daily (t-1)
\item X_i_w: The lagged weekly (t-2:t-5)
\item X_i_m: The lagged monthly  (t-6:t-22)
}
}
\description{
Return the frame in order to provide data for regression analysis
}
\examples{
# Realized measures
realized1 <- drvp(data = xts_data) # realized variances
realized2 <- drsvp(data = xts_data) # realized semivariances
realized <- xts::merge.xts(realized1, realized2, join = 'inner')

# Generate a frame to run HAR
frame <- HAR_frame(ts = realized)
}
