% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_sample.R
\name{MCS}
\alias{MCS}
\title{Model confidence set Analysis}
\usage{
MCS(ls = NULL, alpha = NULL)
}
\arguments{
\item{ls}{A list containing the one step ahead forecasts
. Default = NULL.}

\item{alpha}{a scalar in (0,1) indicating the confidence level of the tests.}
}
\value{
A SSM object containing the Superior Set of Models.
}
\description{
Return the Superior Set of Models.
}
\examples{
# Realized measures
realized1 <- drvp(data = xts_data) # realized variances
realized2 <- drsvp(data = xts_data) # realized semivariances
realized <- xts::merge.xts(realized1, realized2, join = 'inner')

# Generate a frame to run HAR
frame <- HAR_frame(ts = realized)

# Rolling window Analysis
forecasts <- rollwin(data = frame, length_training_data = 0.9)

# MCS
MCS <- MCS(ls = forecasts, alpha = 0.05)

}
