% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xts_data}
\alias{xts_data}
\title{xts_data}
\format{
A xts object with variables:
\describe{
\item{last}{last price}
\item{min}{minimum price}
\item{max}{maximum price}
\item{begin}{beginning price}
}
}
\source{
AE Broadcast
}
\usage{
xts_data
}
\description{
Five-minute returns from Ambev SA, B3 SA, Bradesco SA, Notre Dame Intermedica SA,
Itau SA, JBS SA, Magazine Luiza SA, Petrobras SA, Suzano Papel e Celulose SA and
Vale SA spanning the period from July 2018 to January 2021.
}
\keyword{datasets}
